Clarify W3A8 prefill kernel implementation
